+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/ev/EVRefreshDataProvider.kt
@@ -1,8 +1,8 @@
 package com.mapbox.navigation.core.ev
 
 import com.mapbox.api.directions.v5.models.RouteOptions
-import com.mapbox.navigation.core.routeoptions.KEY_ENGINE
-import com.mapbox.navigation.core.routeoptions.VALUE_ELECTRIC
+import com.mapbox.navigation.core.routeoptions.ROUTE_OPTIONS_KEY_ENGINE
+import com.mapbox.navigation.core.routeoptions.ROUTE_OPTIONS_VALUE_ELECTRIC
 import com.mapbox.navigation.core.routeoptions.isEVRoute
 
 internal class EVRefreshDataProvider(
@@ -12,7 +12,7 @@ internal class EVRefreshDataProvider(
     fun get(initialRouteOptions: RouteOptions): Map<String, String> {
         val result = hashMapOf<String, String>()
         if (initialRouteOptions.isEVRoute()) {
-            result[KEY_ENGINE] = VALUE_ELECTRIC
+            result[ROUTE_OPTIONS_KEY_ENGINE] = ROUTE_OPTIONS_VALUE_ELECTRIC
             result.putAll(
                 evDynamicDataHolder.currentData(initialRouteOptions.unrecognizedJsonProperties!!)
             )