+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/reroute/MapboxRerouteOptionsAdapter.kt
@@ -3,19 +3,19 @@ package com.mapbox.navigation.core.reroute
 import androidx.annotation.VisibleForTesting
 import com.mapbox.api.directions.v5.models.RouteOptions
 import com.mapbox.navigation.core.ev.EVDynamicDataHolder
-import com.mapbox.navigation.core.ev.EVRouteOptionsModifier
+import com.mapbox.navigation.core.ev.EVRerouteOptionsAdapter
 
 internal class MapboxRerouteOptionsAdapter @VisibleForTesting constructor(
-    private val optionsModifiers: List<RouteOptionsModifier>
+    private val optionsAdapters: List<RerouteOptionsAdapter>
 ) : RerouteOptionsAdapter {
 
     constructor(evDynamicDataHolder: EVDynamicDataHolder) : this(
-        listOf(EVRouteOptionsModifier(evDynamicDataHolder))
+        listOf(EVRerouteOptionsAdapter(evDynamicDataHolder))
     )
 
     override fun onRouteOptions(routeOptions: RouteOptions): RouteOptions {
-        return optionsModifiers.fold(routeOptions) { value, modifier ->
-            modifier.modify(value)
+        return optionsAdapters.fold(routeOptions) { value, modifier ->
+            modifier.onRouteOptions(value)
         }
     }
 }