+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/reroute/RouteOptionsModifier.kt
@@ -1,8 +1,0 @@
-package com.mapbox.navigation.core.reroute
-
-import com.mapbox.api.directions.v5.models.RouteOptions
-
-internal interface RouteOptionsModifier {
-
-    fun modify(options: RouteOptions): RouteOptions
-}